+++ test/fork/actions/SwapActions.sol
@@ -105,7 +105,7 @@ contract SwapActions is StdCheats {
      */
 
     ITradingLimits.Config memory limitConfig = ctx.tradingLimitsConfig(from);
-    console.log(unicode"🏷️ [%d] Swap until L0In=%d, L0Out=%d on inflow", block.timestamp, uint256(limitConfig.limit0In), uint256(limitConfig.limit0Out));
+    console.log(unicode"🏷️ [%d] Swap until L0In=%d, L0Out=%d on inflow", block.timestamp, uint48(limitConfig.limit0In), uint48(limitConfig.limit0Out));
     uint256 maxPossible;
     uint256 maxPossibleUntilLimit;
     do {
@@ -134,7 +134,7 @@ contract SwapActions is StdCheats {
      */
     ITradingLimits.Config memory limitConfig = ctx.tradingLimitsConfig(from);
     ITradingLimits.State memory limitState = ctx.refreshedTradingLimitsState(from);
-    console.log(unicode"🏷️ [%d] Swap until L1In=%d, L1Out=%d on inflow", block.timestamp, uint256(limitConfig.limit1In), uint256(limitConfig.limit1Out));
+    console.log(unicode"🏷️ [%d] Swap until L1In=%d, L1Out=%d on inflow", block.timestamp, uint48(limitConfig.limit1In), uint48(limitConfig.limit1Out));
     int48 maxPerSwap = limitState.getLimit0FromNetflow0(limitConfig);
     while (limitState.netflow1 + maxPerSwap <= limitState.getLimit1FromNetflow1(limitConfig)) {
       skip(limitConfig.timestep0 + 1);
@@ -199,7 +199,7 @@ contract SwapActions is StdCheats {
      */
 
     ITradingLimits.Config memory limitConfig = ctx.tradingLimitsConfig(to);
-    console.log(unicode"🏷️ [%d] Swap until L0In=%d, L0Out=%d on outflow", block.timestamp, uint256(limitConfig.limit0In), uint256(limitConfig.limit0Out));
+    console.log(unicode"🏷️ [%d] Swap until L0In=%d, L0Out=%d on outflow", block.timestamp, uint48(limitConfig.limit0In), uint48(limitConfig.limit0Out));
     uint256 maxPossible;
     uint256 maxPossibleUntilLimit;
     do {
